chore: move islands to folders
