+++ components/base/Aside.tsx
@@ -1,4 +1,4 @@
-import ThemeSwitcher from "../../islands/ThemeSwitcher.tsx";
+import ThemeSwitcher from "../../islands/misc/ThemeSwitcher.tsx";
 
 //? Renders the sidebar responsible for switching themes
 export function Aside() {
