+++ routes/projects/food-order/success.tsx
@@ -9,7 +9,7 @@ import { StyledHeader } from "../../../components/UI/StyledHeader.tsx";
 //? Delivery truck SVG
 import { DeliveryIcon } from "../../../assets/DeliveryIcon.tsx";
 //? Delivery timer countdown
-import DigitalTimer from "../../../islands/DigitalTimer.tsx";
+import DigitalTimer from "../../../islands/UI/DigitalTimer.tsx";
 //? Middleware responsible for taking the query from the URL
 import { foodOrderSuccessMiddleware } from "../../../middleware/projects/__food-order-success.tsx";
